tests: move Qt tests into separate test_qt.py file
